+++ jkmod/game/jk_client.c
@@ -122,7 +122,10 @@ void JKMod_ClientBegin(int clientNum, qboolean allowTeamReset)
 	trap_GetUserinfo(clientNum, userinfo, sizeof(userinfo));
 
 	// Set default dimension and refresh
-	if (client->ps.stats[JK_DIMENSION] != DIMENSION_FREE) JKMod_DimensionSet(ent, DIMENSION_FREE);
+	if (jkcvar_altDimensionBase.integer) 
+		JKMod_DimensionSet(ent, jkcvar_altDimensionBase.integer);
+	else if (client->ps.stats[JK_DIMENSION] != DIMENSION_FREE) 
+		JKMod_DimensionSet(ent, DIMENSION_FREE);
 
 	// Set player movements
 	if (jkcvar_playerMovement.integer) JKMod_PlayerMovementCheck(ent);