+++ jkmod/game/jk_local.h
@@ -334,6 +334,7 @@ extern	vmCvar_t					jkcvar_emotesPunchDamage;
 extern	vmCvar_t					jkcvar_gamePlay;
 extern	vmCvar_t					jkcvar_gameTypeConfig;
 extern	vmCvar_t					jkcvar_altDimension;
+extern	vmCvar_t					jkcvar_altDimensionBase;
 extern	vmCvar_t					jkcvar_altDimensionTime;
 extern	vmCvar_t					jkcvar_randomBegin;
 extern	vmCvar_t					jkcvar_serverNews;